@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.Toast;
@@ -53,6 +53,7 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent result) {
+        super.onActivityResult(requestCode, resultCode, result);
         if (requestCode == Crop.REQUEST_PICK && resultCode == RESULT_OK) {
             try {
                 MainActivity.select = MediaStore.Images.Media.getBitmap(getContentResolver(), result.getData());
@@ -66,7 +67,7 @@ public class MainActivity extends AppCompatActivity {
 
     public void btnCrop(View view) {
         if (select != null)
-            startActivity(new Intent(this, RoundRetActivity.class));
+            startActivity(new Intent(this, RoundRet.class));
         else
             Toast.makeText(getApplicationContext(), "请先选择图片", Toast.LENGTH_SHORT).show();
     }
